### 2024-07-19 #7015 @editorialbot
[PRE REVIEW]: ProbNumDiffEq.jl: Probabilistic Numerical Solvers for Ordinary Differential Equations in Julia
**Submitting author:** @nathanaelbosch (<a href="http://orcid.org/0000-0003-0139-4622">Nathanael Bosch</a>)
**Repository:** https://github.com/nathanaelbosch/ProbNumDiffEq.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.16.0
