### 2024-07-19 #7015 @jbytecode
@danielskatz - sure! thank you for inviting me, indeed, I like to edit Julia submissions and the subject perfectly fits my experiences.
