### 2024-07-19 #7015 @jbytecode
@nathanaelbosch - Hi, thank you for submitting your work to JOSS. I am the handling editor. We will be doing whole the stuff together from start to finish. Firstly, could you please try to find the DOI information for the missing ones of the bibtex entries? For example the submission

`DifferentialEquations.jl – A Performant and Featur...`

is resolved by using the DOI information

https://doi.org/10.5334/jors.151

and hopefully you can investigate the others. Addition to this, it seems you forgot to cite the Julia programming language. You can use the bibtex entry

```
@article{julia,
  doi       = {10.1137/141000671},
  url       = {https://doi.org/10.1137%2F141000671},
  year      = 2017,
  month     = {jan},
  publisher = {Society for Industrial {\&} Applied Mathematics ({SIAM})},
  volume    = {59},
  number    = {1},
  pages     = {65--98},
  author    = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
  title     = {Julia: A Fresh Approach to Numerical Computing},
  journal   = {{SIAM} Review}
}
```

for Julia, in manuscript, possibly where the first use is located.

The editorial bot suggests us some potential reviewers for the submission, it lists top 10 with highest similarity scores. I will try to invite them while you are editing your bibtex.

Thank you in advance!
