### 2024-07-19 0bc0c7a
Add Julia citation
