+++ paper/paper.bib
@@ -297,3 +297,17 @@
   pages =        {1--29},
   url =          {http://jmlr.org/papers/v25/20-1423.html}
 }
+
+@article{julia,
+  doi       = {10.1137/141000671},
+  url       = {https://doi.org/10.1137%2F141000671},
+  year      = 2017,
+  month     = {jan},
+  publisher = {Society for Industrial {\&} Applied Mathematics ({SIAM})},
+  volume    = {59},
+  number    = {1},
+  pages     = {65--98},
+  author    = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
+  title     = {Julia: A Fresh Approach to Numerical Computing},
+  journal   = {{SIAM} Review}
+}
