### 2024-07-19 cd08886
Add DifferentialEquations.jl DOI
