+++ paper/paper.bib
@@ -90,7 +90,9 @@
   volume =       5,
   number =       1,
   year =         2017,
-  publisher =    {Ubiquity Press}
+  publisher =    {Ubiquity Press},
+  doi =          "10.5334/jors.151",
+  url =          "https://doi.org/10.5334/jors.151",
 }
 
 @misc{probdiffeq,
