### 2024-07-19 31e0a54
Add more DOIs and reformat bibtex
