+++ paper/paper.bib
@@ -41,7 +41,9 @@
   booktitle =    {Thirty-seventh Conference on Neural Information Processing
                   Systems},
   year =         2023,
-  url =          {https://openreview.net/forum?id=2dx5MNs2Ip}
+  url =          {https://openreview.net/forum?id=2dx5MNs2Ip},
+  doi =          "10.5555/3666122.3667881",
+  url =          "https://dl.acm.org/doi/10.5555/3666122.3667881",
 }
 
 @inproceedings{beck2024diffusion,
@@ -64,22 +66,26 @@
   year =         2021,
   eprint =       {2112.02100},
   archivePrefix ={arXiv},
-  primaryClass = {cs.MS}
+  primaryClass = {cs.MS},
+  doi =          "10.48550/arXiv.2112.02100",
+  url =          "https://doi.org/10.48550/arXiv.2112.02100",
 }
 
 @InProceedings{dalton2024,
-  title = 	 {Data-Adaptive Probabilistic Likelihood Approximation for Ordinary Differential Equations},
+  title =        {Data-Adaptive Probabilistic Likelihood Approximation for
+                  Ordinary Differential Equations},
   author =       {Wu, Mohan and Lysy, Martin},
-  booktitle = 	 {Proceedings of The 27th International Conference on Artificial Intelligence and Statistics},
-  pages = 	 {1018--1026},
-  year = 	 {2024},
-  editor = 	 {Dasgupta, Sanjoy and Mandt, Stephan and Li, Yingzhen},
-  volume = 	 {238},
-  series = 	 {Proceedings of Machine Learning Research},
-  month = 	 {02--04 May},
+  booktitle =    {Proceedings of The 27th International Conference on Artificial
+                  Intelligence and Statistics},
+  pages =        {1018--1026},
+  year =         2024,
+  editor =       {Dasgupta, Sanjoy and Mandt, Stephan and Li, Yingzhen},
+  volume =       238,
+  series =       {Proceedings of Machine Learning Research},
+  month =        {02--04 May},
   publisher =    {PMLR},
-  pdf = 	 {https://proceedings.mlr.press/v238/wu24b/wu24b.pdf},
-  url = 	 {https://proceedings.mlr.press/v238/wu24b.html},
+  pdf =          {https://proceedings.mlr.press/v238/wu24b/wu24b.pdf},
+  url =          {https://proceedings.mlr.press/v238/wu24b.html},
 }
 
 @article{rackauckas2017differentialequations,
@@ -96,12 +102,13 @@
 }
 
 @misc{probdiffeq,
-  author = {Kr{\"a}mer, Nicholas},
-  title = {probdiffeq: Probabilistic solvers for differential equations in JAX},
-  year = {2023},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/pnkraemer/probdiffeq}
+  author =       {Kr{\"a}mer, Nicholas},
+  title =        {probdiffeq: Probabilistic solvers for differential equations
+                  in JAX},
+  year =         2023,
+  publisher =    {GitHub},
+  journal =      {GitHub repository},
+  url =          {https://github.com/pnkraemer/probdiffeq}
 }
 
 @misc{jax2018github,
@@ -185,7 +192,9 @@
   year =         2023,
   eprint =       {2310.01145},
   archivePrefix ={arXiv},
-  primaryClass = {math.NA}
+  primaryClass = {math.NA},
+  doi =          {10.48550/arXiv.2310.01145},
+  url =          {https://doi.org/10.48550/arXiv.2310.01145}
 }
 
 @InProceedings{kersting20invprob,
@@ -220,7 +229,9 @@
   url =
[…]
