### 2024-07-19 #7015 @nathanaelbosch
Hello @jbytecode, thank you for taking the time to edit this submission and for supporting me in the process!

I added the DOI for the DifferentialEquations.jl paper and a citation to the Julia paper, as well as some of the other missing DOIs. I could however not find DOIs for the any paper published by PMLR or JMLR. Do you have any suggestion for how I should proceed to find the missing DOIs, or could it also simply be that these publishers do not provide a DOI?

Thank you again for your support!
