### 2024-07-19 #7015 @jbytecode
@nathanaelbosch - Thank you, let me take a look at the changes. If any improvements are required for this issue, we can fix them after the review as well.
