### 2024-07-19 #7015 @jbytecode
@nathanaelbosch - you can use the https://www.crossref.org/ tool to search DOIs (if exist). You can trigger our editorial bot to check the references by typing

`@editorialbot check references`
