### 2024-07-20 6f10bf2
Remove invalid DOIs
