+++ paper/paper.bib
@@ -42,8 +42,6 @@
                   Systems},
   year =         2023,
   url =          {https://openreview.net/forum?id=2dx5MNs2Ip},
-  doi =          "10.5555/3666122.3667881",
-  url =          "https://dl.acm.org/doi/10.5555/3666122.3667881",
 }
 
 @inproceedings{beck2024diffusion,
@@ -230,8 +228,6 @@
                   {https://proceedings.neurips.cc/paper/2021/file/6734fa703f6633ab896eecbdfad8953a-Paper.pdf},
   volume =       34,
   year =         2021,
-  doi =          {10.5555/3540261.3541208},
-  url =          {https://dl.acm.org/doi/10.5555/3540261.3541208},
 }
 
 @inproceedings{kraemer202bvp,
@@ -246,8 +242,6 @@
                   {https://papers.nips.cc/paper/2021/hash/5ca3e9b122f61f8f06494c97b1afccf3-Abstract.html},
   volume =       34,
   year =         2021,
-  doi =          {10.5555/3540261.3541114},
-  url =          {https://dl.acm.org/doi/10.5555/3540261.3541114},
 }
 
 @InProceedings{kraemer22mol,
