### 2024-07-20 #7015 @nathanaelbosch
@jbytecode I removed the invalid DOIs again and unfortunately I was not able to find DOIs for the remaining papers (I suspect that these ML conferences simply do not attribute a DOI to each paper).
