### 2024-07-20 #7015 @jbytecode
@nathanaelbosch - editorial bot commands should place at the start of the posts.

That's okay to leave conference items in bibtex as they were.
