### 2024-07-22 8304b77
CompatHelper: bump compat for TaylorSeries to 0.18, (keep existing compat)
