### 2024-07-22 #7015 @jbytecode
@pnavaro, @nluetts - Thank you for the reply.

@nathanaelbosch - Do you have any suggestions for the potential reviewers? The list (https://reviewers.joss.theoj.org/lookup) contains a list of potential reviewers that are willing to review in JOSS. If so, you can mention their GitHub handles without using the `@`  character to avoid unwanted notifications.

Thank you in advance!
