### 2024-07-23 #7015 @nathanaelbosch
@jbytecode I do not have any direct recommendation in mind, but filtering for "Julia" as a programming language and "differential equations" as keyword see 16 reviewers that could be suited. These also include some names I know from the SciML ecosystem in Julia, such as `avik-pal`, `ranocha`, or `ChrisRackauckas`, but I suppose this could both qualify then as being very well-suited topic-wise, but potentially also being too close to the project. I also found `matt-graham` who might be well-suited as they list both "numerical simulation" and also more probabilistic topics as their topc areas.
