### 2024-07-23 #7015 @jbytecode
👋👋👋 Dear @avik-pal 👋👋👋

I am writing here to you because you have recently published your work in JOSS which is about Julia and computational methods.

[ ProbNumDiffEq.jl: Probabilistic Numerical Solvers for Ordinary Differential Equations in Julia](https://github.com/openjournals/joss-reviews/issues/7015#top)

You can find more information at the top of this Github issue https://github.com/openjournals/joss-reviews/issues/7015

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be about 20 check items for each single reviewer.

Thank you in advance!
