### 2024-08-03 a160f37
CompatHelper: bump compat for TaylorIntegration to 0.16, (keep existing compat) (#321)
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
