+++ Project.toml
@@ -73,7 +73,7 @@ SimpleUnPack = "1"
 StaticArrayInterface = "1.3"
 Statistics = "1"
 StructArrays = "0.4, 0.5, 0.6"
-TaylorIntegration = "0.8, 0.9, 0.10, 0.11, 0.12, 0.13, 0.14, 0.15"
+TaylorIntegration = "0.8, 0.9, 0.10, 0.11, 0.12, 0.13, 0.14, 0.15, 0.16"
 TaylorSeries = "0.10, 0.11, 0.12, 0.13, 0.14, 0.15, 0.16, 0.17, 0.18"
 Test = "1"
 ToeplitzMatrices = "0.7, 0.8"
