### 2024-08-17 #7048 @ranocha
> @PieterjanRobbe, @ranocha - Could you please update your status and inform us how is your review going?
>
> @ranocha - It seems you haven't created your checklist yet. May I kindly ask you to start your review, please?
>
> Thank you in advance.

As I said in https://github.com/openjournals/joss-reviews/issues/7015#issuecomment-2254372560, it will take a few weeks. I didn't see any reason to generate the review checklist before having the time to look at the contribution but I will generate it right now since you seem to prefer this approach.
