+++ src/alg_utils.jl
@@ -1,47 +1,47 @@
 ############################################################################################
-# For the equivalent parts in OrdinaryDiffEq.jl, see:
-# https://github.com/SciML/OrdinaryDiffEq.jl/blob/master/src/alg_utils.jl
+# For the equivalent parts in OrdinaryDiffEqCore.jl, see:
+# https://github.com/SciML/OrdinaryDiffEqCore.jl/blob/master/src/alg_utils.jl
 ############################################################################################
 
-OrdinaryDiffEq._alg_autodiff(::AbstractEK) = Val{true}()
-OrdinaryDiffEq.standardtag(::AbstractEK) = false
-OrdinaryDiffEq.concrete_jac(::AbstractEK) = nothing
+OrdinaryDiffEqDifferentiation._alg_autodiff(::AbstractEK) = Val{true}()
+OrdinaryDiffEqDifferentiation.standardtag(::AbstractEK) = false
+OrdinaryDiffEqDifferentiation.concrete_jac(::AbstractEK) = nothing
 
 @inline DiffEqBase.get_tmp_cache(integ, alg::AbstractEK, cache::AbstractODEFilterCache) =
     (cache.tmp, cache.atmp)
-OrdinaryDiffEq.isfsal(::AbstractEK) = false
+OrdinaryDiffEqCore.isfsal(::AbstractEK) = false
 
 for ALG in [:EK1, :DiagonalEK1]
-    @eval OrdinaryDiffEq._alg_autodiff(::$ALG{CS,AD}) where {CS,AD} = Val{AD}()
-    @eval OrdinaryDiffEq.alg_difftype(::$ALG{CS,AD,DiffType}) where {CS,AD,DiffType} =
+    @eval OrdinaryDiffEqDifferentiation._alg_autodiff(::$ALG{CS,AD}) where {CS,AD} = Val{AD}()
+    @eval OrdinaryDiffEqDifferentiation.alg_difftype(::$ALG{CS,AD,DiffType}) where {CS,AD,DiffType} =
         DiffType
-    @eval OrdinaryDiffEq.standardtag(::$ALG{CS,AD,DiffType,ST}) where {CS,AD,DiffType,ST} =
+    @eval OrdinaryDiffEqDifferentiation.standardtag(::$ALG{CS,AD,DiffType,ST}) where {CS,AD,DiffType,ST} =
         ST
-    @eval OrdinaryDiffEq.concrete_jac(
+    @eval OrdinaryDiffEqDifferentiation.concrete_jac(
         ::$ALG{CS,AD,DiffType,ST,CJ},
     ) where {CS,AD,DiffType,ST,CJ} = CJ
-    @eval OrdinaryDiffEq.get_chunksize(::$ALG{CS}) where {CS} = Val(CS)
-    @eval OrdinaryDiffEq.isimplicit(::$ALG) = true
+    @eval OrdinaryDiffEqDifferentiation.get_chunksize(::$ALG{CS}) where {CS} = Val(CS)
+    @eval OrdinaryDiffEqCore.isimplicit(::$ALG) = true
 end
 
 ############################################
 # Step size control
-OrdinaryDiffEq.isadaptive(::AbstractEK) = true
-OrdinaryDiffEq.alg_order(alg::AbstractEK) = num_derivatives(alg.prior)
-# OrdinaryDiffEq.alg_adaptive_order(alg::AbstractEK) =
+OrdinaryDiffEqCore.isadaptive(::AbstractEK) = true
+OrdinaryDiffEqCore.alg_order(alg::AbstractEK) = num_derivatives(alg.prior)
[…]
