### 2024-08-29 nathanaelbosch/ProbNumDiffEq.jl#323 @PieterjanRobbe
JOSS manuscript review
JOSS manuscript submission: [https://github.com/openjournals/joss-reviews/issues/7048](https://github.com/openjournals/joss-reviews/issues/7048)

Congratulations on a very nice package!

I think the software is well-scoped, aiming at solving ODEs with filtering-based probabilistic numerics. It is nicely embedded in the existing `DifferentialEquations.jl` ecosystem and plays nicely with other Julia packages, e.g., `Optimization.jl` for parameter inference.

The documentation and tutorials seem very well written. The tutorials are straightforward and demonstrate the capabilities of the package.

The paper is also succinct and straightforward. I particularly liked the open and honest comparison to existing alternatives in other languages, e.g., `probdiffeq` in Python/JAX. Also, this package has already been used in a few papers listed in the manuscript.

I do have a couple of (small) suggestions:

(1) The code on the `README.md` doesn't evaluate properly, I think there's a comma missing after the first line in the statement
```julia
plot(sol.t, hcat(stds...)', color=["#CB3C33" "#389826" "#9558B2"]
     label=["std(u1(t))" "std(u2(t))"], xlabel="t", ylabel="standard-deviation")
```
Also, it would be great if this example would actually show the error bars, e.g., by solving the ODE with larger time steps?
(2) Some of the code in the tutorials doesn't work for me:
- In the [Second Order ODEs and Energy Preservation](https://nathanaelbosch.github.io/ProbNumDiffEq.jl/stable/tutorials/dynamical_odes/#Second-Order-ODEs-and-Energy-Preservation) tutorial, the first plot does not show up for me with
```julia
plot(sol, idxs=(3, 4))
```
but it does work with
```julia
plot(sol[:, 3], sol[:, 4])
```
- In that same tutorial, a `using BenchmarkTools` is missing to be able to use the `@btime` macro.
- In the [Solving DAEs with Probabilistic Numerics](https://nathanaelbosch.github.io/ProbNumDiffEq.jl/stable/tutorials/dae/#Solving-DAEs-with-Probabilistic-Numerics) tutorial, I need to add `using LinearAlgebra` to make sure `Diagonal` is defined.
- Similarly, I need to add `using OrdinaryDiffEq` to make sure the `Rodas4` solver is defined.
- I also had issues compiling `ModelingToolkit.jl` on my julia installation (`v1.9.3`), so I couldn't execute the code below the `modelingtoolkitize` function [here](https://nathanaelbosch.github.io/ProbNumDiffEq.jl/stable/tutorials/dae/#Is-index-reduction-still-worth-it?).
