### 2024-08-30 nathanaelbosch/ProbNumDiffEq.jl#324 @nathanaelbosch
Thanks for bringing this up. This is likely related to #322 and unfortunately I didn't have the time to keep up with the OrdinaryDiffEq.jl changes these last weeks. But I will do my best to resolve this as quickly as I can now.
