### 2024-08-30 nathanaelbosch/ProbNumDiffEq.jl#324 @nathanaelbosch
#325 will resolve this. To use the package right away you could install this branch from github with
```
] add ProbNumDiffEq#hotfix/get_fsalfirstlast
```
But I will also make sure to get the PR merged with passing tests asap and then do a quick patch release.
