### 2024-08-30 nathanaelbosch/ProbNumDiffEq.jl#324 @ranocha
Thanks! I will wait for the patch release
