+++ src/caches.jl
@@ -251,7 +251,9 @@ function OrdinaryDiffEq.alg_cache(
     )
 end
 
-get_uf(f, t, p, ::Val{true}) = OrdinaryDiffEq.UJacobianWrapper(f, t, p)
-get_uf(f, t, p, ::Val{false}) = OrdinaryDiffEq.UDerivativeWrapper(f, t, p)
+get_uf(f, t, p, ::Val{true}) =
+    OrdinaryDiffEq.OrdinaryDiffEqDifferentiation.UJacobianWrapper(f, t, p)
+get_uf(f, t, p, ::Val{false}) =
+    OrdinaryDiffEq.OrdinaryDiffEqDifferentiation.UDerivativeWrapper(f, t, p)
 
 OrdinaryDiffEq.get_fsalfirstlast(cache::EKCache, rate_prototype) = nothing, nothing
