### 2024-08-30 e4ea30c
Update to OrdinaryDiffEq split version (#326)
* Update to OrdinaryDiffEq split version

This both updates to the split version (reducing dependencies) and fixes some of the internals changes that occurred.

* The URL in precompile.jl shouldn't get changed

---------

Co-authored-by: Nathanael Bosch <nathanael.bosch@uni-tuebingen.de>
