+++ src/derivative_utils.jl
@@ -7,12 +7,12 @@ function calc_H!(H, integ, cache)
     elseif integ.alg isa EK1
         calc_H_EK0!(H, integ, cache)
         # @assert integ.u == @view x_pred.μ[1:(q+1):end]
-        OrdinaryDiffEq.calc_J!(ddu, integ, cache, true)
+        OrdinaryDiffEqDifferentiation.calc_J!(ddu, integ, cache, true)
         _ddu = size(ddu, 2) != d ? view(ddu, 1:d, :) : ddu
         _matmul!(H, _ddu, cache.SolProj, -1.0, 1.0)
     elseif integ.alg isa DiagonalEK1
         calc_H_EK0!(H, integ, cache)
-        OrdinaryDiffEq.calc_J!(ddu, integ, cache, true)
+        OrdinaryDiffEqDifferentiation.calc_J!(ddu, integ, cache, true)
         _ddu = size(ddu, 2) != d ? view(ddu, 1:d, :) : ddu
         ddu_diag = if size(ddu, 2) == d
             # the normal case: just extract the diagonal
