### 2024-08-30 564495c
Fix the failing plotting example from the README
