### 2024-08-30 nathanaelbosch/ProbNumDiffEq.jl#323 @nathanaelbosch
Thank you very much for taking the time to review the package! I really appreciate your comments and your review and feedback helps a lot. I'll go through your suggestions one by one.

> The code on the README.md doesn't evaluate properly, I think there's a comma missing after the first line in the statement

Good catch! Apparently a comma got lost in the last change. I will fix this asap.

> Also, it would be great if this example would actually show the error bars, e.g., by solving the ODE with larger time steps?

I also had this idea when I wrote the first README and in a previous version this was the case (see [here](https://github.com/nathanaelbosch/ProbNumDiffEq.jl/blob/c5ab0ed71e4347743fcbebe3368f80b27dcd55a6/README.md)). The problem I encountered is that with such a README, people expect to actually _see_ the numerical error in their own examples, but in most cases the numerical error is much smaller than the solution. This lead to confusion when interacting with potential users, both on github and in person. This is why the current README now covers a (hopefully) more representative setting. What do you think, shoudl I keep the current one or rather revert to the earlier version? Depending on your feedback I would also be very happy to adjust the README example to a more informative one.

> In the [Second Order ODEs and Energy Preservation](https://nathanaelbosch.github.io/ProbNumDiffEq.jl/stable/tutorials/dynamical_odes/#Second-Order-ODEs-and-Energy-Preservation) tutorial, the first plot does not show up for me

I just tried it with an up-to-date repository and there the plotting works. I can unfortunately not test the current release due to an installation error (#324) but I will test this again with the upcoming patch release. I will report back once I did this.

> In that same tutorial, a using BenchmarkTools is missing to be able to use the @btime macro.

You are totally right of course! That part of the doc is not Julia code but simply text so CI could not catch this. I will add a `using BenchmarkTools` line.

> In the [Solving DAEs with Probabilistic Numerics](https://nathanaelbosch.github.io/ProbNumDiffEq.jl/stable/tutorials/dae/#Solving-DAEs-with-Probabilistic-Numerics) tutorial, I need to add using LinearAlgebra to make sure Diagonal is defined.

Good catch again! I will add `LinearAlgebra` to the imports.

> Similarly, I need to add using OrdinaryDiffEq to make sure the Rodas4 solver is defined.

OrdinaryDiffEq.jl is part of the imports [here](https://github.com/nathanaelbosch/ProbNumDiffEq.jl/blob/564495c4ad6b7e1f85dc45c4aac9f1df76c16576/docs/src/tutorials/dae.md?plain=1#L15).

> I also had issues compiling ModelingToolkit.jl on my julia installation (v1.9.3), so I couldn't execute the code below the modelingtoolkitize function [here](https://nathanaelbosch.github.io/ProbNumDiffEq.jl/stable/tutorials/dae/#Is-index-reduction-still-worth-it?).

I had no issues precompiling ModelingToolkit.jl with Julia 1.10.5, and also with 1.9.3. Though I also recommend to use ProbNumDiffEq.jl with Julia 1.10 and I set the compat entry accordingly in the [Project.toml](https://github.com/nathanaelbosch/ProbNumDiffEq.jl/blob/564495c4ad6b7e1f85dc45c4aac9f1df76c16576/Project.toml#L86C1-L86C15), as I do not test for older Julia versions, so I cannot guarantee that the code runs for older Julia versions.

Thanks again for all the helpful comments and for taking the time to go through the package and documentation so thoroughly! Let me know if anything else comes up.
