### 2024-08-30 nathanaelbosch/ProbNumDiffEq.jl#323 @nathanaelbosch
I just installed Julia 1.9.3 and I saw that I can not resolve the package dependencies on the current `main` branch, so it seems the most up-to-date version will not be compatible with Julia versions below 1.10 anymore. So I suppose that the environment in which you tested the code also used older package versions. Which is also a good thing, as you did not run into the current installation issue (#324), but I suppose that this also explains the failures in the plotting code possibly the ModelingToolkig compilation error, and I would expect that an update to Julia 1.10 resolves these. Does this address your issues to a satisfactory level? If not please let me know how I could resolve these best. Thanks again!
