### 2024-08-30 nathanaelbosch/ProbNumDiffEq.jl#324 @nathanaelbosch
I just released 0.16.1 and checked that I can indeed install it in a fresh environment. I will close this for now, but if you still encounter some installation problem please feel free to reopen. And of course thank you a lot for taking the time to review this package!
