### 2024-08-30 nathanaelbosch/ProbNumDiffEq.jl#323 @PieterjanRobbe
Hi @nathanaelbosch, thanks for the quick response!

I tried again with a fresh install of 1.10.5, and it seems like indeed the plotting issue as well as the modelingtoolkit issue are now resolved, thanks!

As for the README, I understand your reasoning. I think you could just leave it as it is? I'll mark my review as complete.
