### 2024-08-30 nathanaelbosch/ProbNumDiffEq.jl#323 @nathanaelbosch
It's good to hear everything has been resolved. Thanks again for taking the time to review the paper and package!
