### 2024-09-21 nathanaelbosch/ProbNumDiffEq.jl#324 @nathanaelbosch
Hi @ranocha, were you able to confirm that the package can now be installed correctly? If there is any other related open issue, please let me know and I'll do my best to take care of it.
