### 2024-09-23 nathanaelbosch/ProbNumDiffEq.jl#324 @ranocha
Thanks, it works for me indeed
