### 2024-09-23 nathanaelbosch/ProbNumDiffEq.jl#327 @ranocha
JOSS review
This is part of the review https://github.com/openjournals/joss-reviews/issues/7048

Thanks a lot for your contributions 🙂 Please find below some comments based on the review checklist:

- [ ] The performance claims "ProbNum (Wenger et al., 2021) is a Python package that provides probabilistic numerical solvers for ODEs [...] with its reliance on Python and numpy, it is also generally less performant." and "ProbDiffEq provides ODE solvers with similar performance as those implemented in ProbNumDiffEq.jl." are not backed up by data in the paper itself.
- [ ] It's nice that you mention related packages in the README. Could you please briefly point out how your package differs from theirs so that people know whether they should look deeper into your package or the alternatives?
- [ ] Some citation references in the docstrings do not seem to work properly, e.g.,
    ```julia
    help?> EK1
    search: EK1 DiagonalEK1

    EK1(; order=3,
            smooth=true,
            prior=IWP(order),
            diffusionmodel=DynamicDiffusion(),
            initialization=TaylorModeInit(num_derivatives(prior)),
            kwargs...)

    Gaussian ODE filter with first-order vector field linearization.

    This is a semi-implicit, L-stable ODE solver so it can handle stiffness quite well
    tronarp18probsol (@cite), and it generally produces more expressive posterior covariances than the
    EK0. However, as typical implicit ODE solvers it scales cubically with the ODE dimension
    krämer21highdim (@cite), so if you're solving a high-dimensional non-stiff problem you might want
    to give the EK0 a try.
    ```
- [ ] The benchmark in https://nathanaelbosch.github.io/ProbNumDiffEq.jl/stable/tutorials/dynamical_odes/#Benchmark:-Solving-second-order-ODEs-is-*faster* uses also different arguments `EK1(order=3)` vs. `EK1(order=4)`, not only `prob` vs. `prob2`. Could you please comment on that in the tutorial?
