### 2024-09-23 ad3faa6
Update and correct second-order-ODE docs
