### 2024-09-24 nathanaelbosch/ProbNumDiffEq.jl#327 @ranocha
> Is this sufficient evidence to keep the part in the paper as it is, or should I soften the statement or add a sentence to support it?

I would be fine if you explained this a bit like above. Maybe you can cite some of the benchmarks of OrdinaryDiffEq.jl etc.

> Does this add the desired information and resolve your concern?

👍 Please remember to merge the PR and update the paper branch accordingly.

> Personally I would prefer to rather focus on the online documentation, at the cost of some imperfections in the REPL, but I would be very happy about your input here. What would be the best way to resolve this issue to a level that is satisfactory to you?

That sounds fine. Please open an issue in your repo so that people know that this is a known issue (linking to the issue you mentioned above).

> Good catch! I updated the code to use order=3 in both benchmarks.

👍

> Is this satisfactory to resolve this part of the issue?

Ok, let's keep it as it is.

Thank you for your contribution to the open-source software ecosystem. Feel free to close this issue once you have merged the corresponding PR.
