### 2024-09-24 8ef8f52
Improvements based on review (#328)
* README: more explanation for related-package trade-offs

* Update and correct second-order-ODE docs

* Update the related-package section in `docs/src/index.md`
