### 2024-09-24 d615558
More details on the performance comparison to ProbNum and ProbDiffEq
