### 2024-09-24 nathanaelbosch/ProbNumDiffEq.jl#327 @nathanaelbosch
Thank you for the fast and helpful reply!

I updated the text in the paper and in particular I mentioned that both packages include benchmarks that compare to scipy, as those are currently the main evidence for the "similar performance" claim.
I also merged the PR and rebased the paper branch on main, and I opened #329 to keep track of the rendering issues of citations in the REPL.

I believe this should resolve the remaining open points, and I will therefore close this issue (but of course if you have more comments please just re-open it). Thank you again for taking the time to review the package and for providing such helpful comments!
