### 2024-09-25 #7048 @ranocha
From my point of view, the paper can be accepted for publication. Thanks a lot for your contribution to the open-source ecosystem, @nathanaelbosch
