### 2024-09-26 #7048 @jbytecode
@nathanaelbosch - I have just sent a pull request regarding to some minor fixes in both the manuscript and the bibtex.

Please review the changes. If you're agreed with them, please merge the PR.

Thank you in advance.
