### 2024-09-27 e8ae52d
Update paper/paper.md
