### 2024-09-27 02805e2
Minor fix on paper and bibtex (#330)
* update bibtex

* update paper.md

* Update paper/paper.md

---------

Co-authored-by: Nathanael Bosch <nathanael.bosch@uni-tuebingen.de>
