+++ paper/paper.bib
@@ -56,7 +56,7 @@
 }
 
 @misc{wenger2021probnum,
-  title =        {ProbNum: Probabilistic Numerics in Python},
+  title =        {ProbNum: Probabilistic Numerics in {P}ython},
   author =       {Jonathan Wenger and Nicholas Krämer and Marvin Pförtner and
                   Jonathan Schmidt and Nathanael Bosch and Nina Effenberger and
                   Johannes Zenn and Alexandra Gessner and Toni Karvonen and
@@ -87,8 +87,8 @@
 }
 
 @article{rackauckas2017differentialequations,
-  title =        {{DifferentialEquations.jl} – A Performant and Feature-Rich
-                  Ecosystem for Solving Differential Equations in Julia},
+  title =        {{DifferentialEquations.jl} – {A} Performant and Feature-Rich
+                  Ecosystem for Solving Differential Equations in {J}ulia},
   author =       {Rackauckas, Christopher and Nie, Qing},
   journal =      {Journal of Open Research Software},
   volume =       5,
@@ -155,7 +155,7 @@
 
 @article{kersting20,
   author =       {Kersting, Hans and Sullivan, T. J. and Hennig, Philipp},
-  title =        {Convergence rates of Gaussian ODE filters},
+  title =        {Convergence rates of {G}aussian ODE filters},
   journal =      {Statistics and Computing},
   year =         2020,
   month =        {Nov},
