### 2024-09-27 #7048 @nathanaelbosch
Thanks a lot for the pull request, and for taking the time to edit this submission!

I checked the ORCID and made a release, and created the following entry on zenodo: https://zenodo.org/records/13851678. Please let me know if there is anything else I should do or adjust on github or on zenodo.
