### 2024-09-28 f300dee
CompatHelper: bump compat for DiffEqCallbacks to 4, (keep existing compat)
