### 2024-09-28 #7048 @jbytecode
@nathanaelbosch - We are almost done. The whole paper and bibtex seem good to me. Please take a full proof read one more time and fix any typos or wording in the manuscript. In this state, please do not make major changes. Ping me when you've finished your latest controls.
