### 2024-09-28 0bc7944
Minor improvements to wording and commas in the paper
