### 2024-09-28 6d4ca4a
CompatHelper: bump compat for DiffEqCallbacks to 4 and drop 2.3.6 (#331)
* CompatHelper: bump compat for DiffEqCallbacks to 4, (keep existing compat)

* Update Project.toml

---------

Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
Co-authored-by: Nathanael Bosch <nathanael.bosch@uni-tuebingen.de>
