+++ Project.toml
@@ -51,7 +51,7 @@ RecipesBaseExt = "RecipesBase"
 ArrayAllocators = "0.3"
 BlockArrays = "1"
 DiffEqBase = "6.122"
-DiffEqCallbacks = "2.36, 3"
+DiffEqCallbacks = "3, 4"
 DiffEqDevTools = "2"
 DocStringExtensions = "0.9"
 FastBroadcast = "0.2, 0.3"
