### 2024-09-28 #7048 @nathanaelbosch
@jbytecode I went through the paper once more and made a few small adjustments in this commit: https://github.com/nathanaelbosch/ProbNumDiffEq.jl/commit/0bc7944a7ffe042974bdfb8423318788d615a5d4. With this, from my side the paper looks good to go!

Let me know if there is anything else I should do. And thanks once more for taking the time to edit this submission!
