### 2024-09-28 #7048 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
